requirements.txt
# alembic==1.17.2
# blinker==1.9.0
# cachetools==6.2.4
# certifi==2025.11.12
# charset-normalizer==3.4.4
# click==8.3.1
# dnspython==2.8.0
# email-validator==2.3.0
# Flask==3.1.2
# Flask-Login==0.6.3
# Flask-Migrate==4.1.0
# Flask-SQLAlchemy==3.1.1
# Flask-WTF==1.2.2
# git-filter-repo==2.47.0
# google-api-core==2.28.1
# google-api-python-client==2.187.0
# google-auth==2.41.1
# google-auth-httplib2==0.3.0
# google-auth-oauthlib==1.2.3
# googleapis-common-protos==1.72.0
# greenlet==3.3.0
# gunicorn==23.0.0
# httplib2==0.31.0
# idna==3.11
# itsdangerous==2.2.0
# Jinja2==3.1.6
# Mako==1.3.10
# MarkupSafe==3.0.3
# oauthlib==3.3.1
# packaging==25.0
# proto-plus==1.27.0
# protobuf==6.33.2
# pyasn1==0.6.1
# pyasn1_modules==0.4.2
# pyparsing==3.2.5
# python-dotenv==1.2.1
# requests==2.32.5
# requests-oauthlib==2.0.0
# rsa==4.9.1
# SQLAlchemy==2.0.45
# typing_extensions==4.15.0
# uritemplate==4.2.0
# urllib3==2.6.2
# Werkzeug==3.1.4
# WTForms==3.2.1



Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.1.4
gunicorn==23.0.0
psycopg2-binary==2.9.6
google-auth==2.41.1
google-api-python-client==2.187.0
requests==2.32.5
python-dotenv==1.2.1